requirements.txt
customtkinter
requests
python-docx
PyMuPDF
pytesseract